+++ src/identity/punish.rs
@@ -129,7 +129,7 @@ impl IdentityService {
             proof_id,
             timestamp,
         };
-        self.penalties.insert_moderator_penalty(user, event).await
+        self.penalties.set_moderator_penalty(user, event).await
     }
 
     pub async fn punish_for_forgetting_with_timestamp(
@@ -147,7 +147,7 @@ impl IdentityService {
             timestamp,
         };
         self.penalties
-            .insert_forgotten_penalty(user, vouchee, event)
+            .set_forgotten_penalty(user, vouchee, event)
             .await
     }
